agents-swarm/stellar_immigration/requirements.txt
functions-framework==3.*
google-cloud-aiplatform
python-dotenv
supabase
python-dateutil
